GROUP BY non-aggregate expressions in the SELECT-list
Quote:

SQL92 states that any non-aggregate expressions appearing in the
SELECT-list must appear in the GROUP BY-clause; SQL99 and on state
that any non-aggregate expressions that appear in the SELECT-list
must be functionally dependent upon the list of expressions appearing
in the GROUP BY. In this context, "functionally dependent" simply
means that for each unique combination of values returned by the
expressions that make up the GROUP BY-clause, the non-aggregate
expression necessarily yields exactly one value.

This patch is an initial implementation of SQL99 behavior.